requirements.txt
pymysql
python-dotenv
requests